+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [CalVer](https://calver.org/).
 ## [Unreleased]
 ### Added
 - Filter Icon added in layer tree next to the eye icon. Indicates the present of a filter on the current layer 
+- The "Sign out" in the Session dialog,  deletes the autoconnect.gc2 cookie. 
 
 
 ## [2023.6.0] - 2023-20-6